+++ package.json
@@ -7,8 +7,16 @@
   "author": {
     "name": "Cameron J Roe"
   },
+  "babel": {
+    "presets": [
+      "es2015"
+    ]
+  },
   "scripts": {
-    "dev": "grunt dev"  
+    "dev": "grunt dev",
+    "webpack": "webpack",
+    "build": "webpack -p --config webpack.prod.babel.js",
+    "server": "webpack-dev-server --content-base ./demo --hot --inline --no-info"
   },
   "repository": {
     "type": "git",
@@ -22,23 +30,39 @@
     }
   ],
   "dependencies": {
-    "jquery": "*"
+    "jquery": "^2.1.4"
   },
   "devDependencies": {
+    "autoprefixer": "^6.2.2",
+    "babel-core": "^6.3.26",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.3.13",
+    "css-loader": "^0.23.1",
+    "cssnano": "^3.4.0",
+    "exports-loader": "^0.6.2",
+    "expose-loader": "^0.7.1",
+    "extract-text-webpack-plugin": "^0.9.1",
     "grunt": "^0.4.5",
     "grunt-autoprefixer": "^3.0.0",
-    "grunt-contrib-clean": "~0.4.0",
-    "grunt-contrib-concat": "~0.3.0",
-    "grunt-contrib-connect": "^0.8.0",
-    "grunt-contrib-cssmin": "^0.12.3",
-    "grunt-contrib-jasmine": "^0.7.0",
-    "grunt-contrib-jshint": "~0.10.0",
-    "grunt-contrib-uglify": "~0.2.0",
-    "grunt-contrib-watch": "~0.4.0",
-    "grunt-sass": "^0.14.0",
+    "grunt-contrib-clean": "~0.7.0",
+    "grunt-contrib-concat": "~0.5.1",
+    "grunt-contrib-connect": "^0.11.2",
+    "grunt-contrib-cssmin": "^0.14.0",
+    "grunt-contrib-jasmine": "^0.9.2",
+    "grunt-contrib-jshint": "~0.11.3",
+    "grunt-contrib-uglify": "~0.11.0",
+    "grunt-contrib-watch": "~0.6.1",
+    "grunt-sass": "^1.1.0",
+    "imports-loader": "^0.6.5",
     "load-grunt-configs": "^0.4.0",
-    "load-grunt-tasks": "^0.6.0",
-    "node-bourbon": "^1.2.3"
+    "load-grunt-tasks": "^3.4.0",
+    "node-bourbon": "^4.2.3",
+    "postcss-loader": "^0.8.0",
+    "sass-loader": "^3.1.2",
+    "script-loader": "^0.6.1",
+    "style-loader": "^0.13.0",
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0"
   },
   "keywords": [
     "jQuery",
