--- src/css/jquery.flipper.css
@@ -1,75 +1,0 @@
-/*
-    jQuery Flipper Demo
-    Author: @cameronjroe
-    Github: https://github.com/cameronjroe/jquery-flipper.git
-*/
-*[data-flipper] {
-  display: none; }
-
-.flipper-container {
-  position: relative; }
-  .flipper-container .flipper-element {
-    position: relative;
-    display: inline-block;
-    width: 100%;
-    height: 100%;
-    -webkit-transform-style: preserve-3d;
-            transform-style: preserve-3d; }
-    .flipper-container .flipper-element ._flipper-front {
-      position: relative;
-      z-index: 2;
-      width: 100%;
-      height: 100%;
-      -webkit-backface-visibility: hidden;
-              backface-visibility: hidden; }
-    .flipper-container .flipper-element ._flipper-back {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      z-index: 0;
-      -webkit-backface-visibility: hidden;
-              backface-visibility: hidden;
-      -webkit-transform: rotate3d(0, 1, 0, 180deg) translateZ(1px);
-              transform: rotate3d(0, 1, 0, 180deg) translateZ(1px); }
-  .flipper-container ._flipper-right.flipper-flipped {
-    -webkit-transform: rotate3d(0, 1, 0, -180deg);
-            transform: rotate3d(0, 1, 0, -180deg); }
-  .flipper-container ._flipper-left.flipper-flipped {
-    -webkit-transform: rotate3d(0, 1, 0, 180deg);
-            transform: rotate3d(0, 1, 0, 180deg); }
-  .flipper-container ._flipper-right-slide {
-    -webkit-transform-origin: center right;
-        -ms-transform-origin: center right;
-            transform-origin: center right; }
-    .flipper-container ._flipper-right-slide.flipper-flipped {
-      -webkit-transform: translateX(-100%) rotate3d(0, 1, 0, -180deg);
-              transform: translateX(-100%) rotate3d(0, 1, 0, -180deg); }
-  .flipper-container ._flipper-left-slide {
-    -webkit-transform-origin: center left;
-        -ms-transform-origin: center left;
-            transform-origin: center left; }
-    .flipper-container ._flipper-left-slide.flipper-flipped {
-      -webkit-transform: translateX(100%) rotate3d(0, 1, 0, 180deg);
-              transform: translateX(100%) rotate3d(0, 1, 0, 180deg); }
-  .flipper-container ._flipper-up {
-    -webkit-transform-origin: center center;
-        -ms-transform-origin: center center;
-            transform-origin: center center; }
-    .flipper-container ._flipper-up > ._flipper-back {
-      -webkit-transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
-              transform: rotate3d(1, 0, 0, 180deg) translateZ(1px); }
-    .flipper-container ._flipper-up.flipper-flipped {
-      -webkit-transform: rotate3d(1, 0, 0, 180deg);
-              transform: rotate3d(1, 0, 0, 180deg); }
-  .flipper-container ._flipper-down {
-    -webkit-transform-origin: center center;
-        -ms-transform-origin: center center;
-            transform-origin: center center; }
-    .flipper-container ._flipper-down > ._flipper-back {
-      -webkit-transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
-              transform: rotate3d(1, 0, 0, 180deg) translateZ(1px); }
-    .flipper-container ._flipper-down.flipper-flipped {
-      -webkit-transform: rotate3d(1, 0, 0, -180deg);
-              transform: rotate3d(1, 0, 0, -180deg); }
