+++ src/scss/jquery.flipper.scss
@@ -1,80 +1,73 @@
-@import "bourbon";
-/*
-    jQuery Flipper Demo
-    Author: @cameronjroe
-    Github: https://github.com/cameronjroe/jquery-flipper.git
-*/
-
 $flipper-el: '.flipper-element';
 
 *[data-flipper] {
-    display: none;
+  display: none;
 }
 
 .flipper-container {
+  position: relative;
+  #{$flipper-el} {
     position: relative;
-    #{$flipper-el} {
-        position: relative;
-        display: inline-block;
-        width: 100%;
-        height: 100%;
-        transform-style: preserve-3d;
-        ._flipper-front {
-            position: relative;
-            z-index: 2;
-            width: 100%;
-            height: 100%;
-            backface-visibility: hidden;
-        }
-        ._flipper-back {
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            z-index: 0;
-            backface-visibility: hidden;
-            transform: rotate3d(0, 1, 0, 180deg) translateZ(1px);
-        }
+    display: inline-block;
+    width: 100%;
+    height: 100%;
+    transform-style: preserve-3d;
+    ._flipper-front {
+      position: relative;
+      z-index: 2;
+      width: 100%;
+      height: 100%;
+      backface-visibility: hidden;
     }
-    ._flipper-right {
-        &.flipper-flipped {
-            transform: rotate3d(0, 1, 0, -180deg);
-        }
+    ._flipper-back {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      z-index: 0;
+      backface-visibility: hidden;
+      transform: rotate3d(0, 1, 0, 180deg) translateZ(1px);
     }
-    ._flipper-left {
-        &.flipper-flipped {
-            transform: rotate3d(0, 1, 0, 180deg);
-        }   
+  }
+  ._flipper-right {
+    &.flipper-flipped {
+      transform: rotate3d(0, 1, 0, -180deg);
     }
-    ._flipper-right-slide {
-        transform-origin: center right;
-        &.flipper-flipped {
-            transform: translateX(-100%) rotate3d(0, 1, 0, -180deg);
-        }
+  }
+  ._flipper-left {
+    &.flipper-flipped {
+      transform: rotate3d(0, 1, 0, 180deg);
     }
-    ._flipper-left-slide {
-        transform-origin: center left;
-        &.flipper-flipped {
-            transform: translateX(100%) rotate3d(0, 1, 0, 180deg);
-        }
+  }
+  ._flipper-right-slide {
+    transform-origin: center right;
+    &.flipper-flipped {
+      transform: translateX(-100%) rotate3d(0, 1, 0, -180deg);
     }
-    ._flipper-up {
-        transform-origin: center center;
-        & > ._flipper-back {
-            transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
-        }
-        &.flipper-flipped {
-            transform: rotate3d(1, 0, 0, 180deg);
-        }
+  }
+  ._flipper-left-slide {
+    transform-origin: center left;
+    &.flipper-flipped {
+      transform: translateX(100%) rotate3d(0, 1, 0, 180deg);
     }
-    ._flipper-down {
-        transform-origin: center center;
-        & > ._flipper-back {
-            transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
-        }
-        &.flipper-flipped {
-            transform: rotate3d(1, 0, 0, -180deg);
-        }
+  }
+  ._flipper-up {
+    transform-origin: center center;
+    & > ._flipper-back {
+      transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
     }
+    &.flipper-flipped {
+      transform: rotate3d(1, 0, 0, 180deg);
+    }
+  }
+  ._flipper-down {
+    transform-origin: center center;
+    & > ._flipper-back {
+      transform: rotate3d(1, 0, 0, 180deg) translateZ(1px);
+    }
+    &.flipper-flipped {
+      transform: rotate3d(1, 0, 0, -180deg);
+    }
+  }
 }
